move all the owl components to llm_thread
