+++ llm_thread/static/src/components/llm_chat_thread_header/llm_chat_thread_header.xml
@@ -51,7 +51,7 @@
         class="o_ThreadDropdowns d-flex align-items-center"
       >
                 <!-- LLM Provider Dropdown -->
-                <div class="dropdown">
+        <div class="dropdown">
                     <button
             class="btn btn-light dropdown-toggle d-flex align-items-center gap-2"
             type="button"
@@ -129,6 +129,46 @@
                         </t>
                     </ul>
                 </div>
+                <!-- Tools Dropdown -->
+            <t t-if="!isSmall">
+                <div class="dropdown ms-2">
+                    <button
+              class="btn btn-light dropdown-toggle d-flex align-items-center gap-2"
+              type="button"
+              data-bs-toggle="dropdown"
+              aria-expanded="false"
+              style="max-width: 200px;"
+            >
+                        Tools
+                        <t t-if="thread.selectedTools.length > 0">
+                            <span class="badge bg-secondary ms-1">
+                                <t t-esc="thread.selectedTools.length" />
+                            </span>
+                        </t>
+                    </button>
+                    <div class="dropdown-menu p-2">
+                        <h6 class="dropdown-header">Select Tools</h6>
+                        <t
+                t-foreach="llmChat.tools"
+                t-as="tool"
+                t-key="tool.id"
+              >
+                            <div class="form-check">
+                                <input
+                    type="checkbox"
+                    class="form-check-input"
+                    t-att-value="tool.id"
+                    t-att-checked="thread.selectedToolIds.includes(tool.id)"
+                    t-on-change="(ev) => onToolSelectChange(ev, tool)"
+                  />
+                                <label class="form-check-label">
+                                    <t t-esc="tool.name" />
+                                </label>
+                            </div>
+                        </t>
+                    </div>
+                </div>
+            </t>
             </div>
 
             <button
