+++ llm_thread/static/src/components/message/message.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8" ?>
 <templates xml:space="preserve">
     <t
-    t-name="llm_tool.Message"
+    t-name="llm_thread.Message"
     t-inherit="mail.Message"
     t-inherit-mode="extension"
     owl="1"
