+++ llm_tool/data/llm_tool_data.xml
@@ -14,7 +14,9 @@
     <record id="llm_tool_odoo_record_creator" model="llm.tool">
         <field name="name">odoo_record_creator</field>
         <field name="implementation">odoo_record_creator</field>
-        <field name="description">Create new records in any Odoo model. This tool allows you to insert data into the database by specifying the model name and a dictionary of field values for the new record. Due to its data modification capabilities, this tool requires user consent before execution.</field>
+        <field
+      name="description"
+    >Create new records in any Odoo model. This tool allows you to insert data into the database by specifying the model name and a dictionary of field values for the new record. Due to its data modification capabilities, this tool requires user consent before execution.</field>
         <field name="default" eval="True" />
         <field name="active" eval="True" />
         <field name="requires_user_consent" eval="True" />
@@ -23,7 +25,9 @@
     <record id="llm_tool_odoo_record_updater" model="llm.tool">
         <field name="name">odoo_record_updater</field>
         <field name="implementation">odoo_record_updater</field>
-        <field name="description">Update existing records in any Odoo model. This tool allows you to modify data in the database by specifying the model name, domain filters to identify records, and a dictionary of field values to update. For safety, it defaults to updating only one record at a time. Due to its data modification capabilities, this tool requires user consent before execution.</field>
+        <field
+      name="description"
+    >Update existing records in any Odoo model. This tool allows you to modify data in the database by specifying the model name, domain filters to identify records, and a dictionary of field values to update. For safety, it defaults to updating only one record at a time. Due to its data modification capabilities, this tool requires user consent before execution.</field>
         <field name="default" eval="True" />
         <field name="active" eval="True" />
         <field name="requires_user_consent" eval="True" />
