+++ llm_tool/models/llm_provider.py
@@ -1,4 +1,3 @@
-import json
 import logging
 
 from odoo import models
