--- llm_tool/static/src/components/llm_chat_thread_header/llm_chat_thread_header.xml
@@ -1,52 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<templates xml:space="preserve">
-    <t
-    t-name="llm_tool.LLMChatThreadHeader"
-    t-inherit="llm_thread.LLMChatThreadHeader"
-    t-inherit-mode="extension"
-    owl="1"
-  >
-    <xpath expr="//div[hasclass('o_ThreadDropdowns')]" position="inside">
-            <!-- Tools Dropdown -->
-            <t t-if="!isSmall">
-                <div class="dropdown ms-2">
-                    <button
-            class="btn btn-light dropdown-toggle d-flex align-items-center gap-2"
-            type="button"
-            data-bs-toggle="dropdown"
-            aria-expanded="false"
-            style="max-width: 200px;"
-          >
-                        Tools
-                        <t t-if="thread.selectedTools.length > 0">
-                            <span class="badge bg-secondary ms-1">
-                                <t t-esc="thread.selectedTools.length" />
-                            </span>
-                        </t>
-                    </button>
-                    <div class="dropdown-menu p-2">
-                        <h6 class="dropdown-header">Select Tools</h6>
-                        <t
-              t-foreach="llmChat.tools"
-              t-as="tool"
-              t-key="tool.id"
-            >
-                            <div class="form-check">
-                                <input
-                  type="checkbox"
-                  class="form-check-input"
-                  t-att-value="tool.id"
-                  t-att-checked="thread.selectedToolIds.includes(tool.id)"
-                  t-on-change="(ev) => onToolSelectChange(ev, tool)"
-                />
-                                <label class="form-check-label">
-                                    <t t-esc="tool.name" />
-                                </label>
-                            </div>
-                        </t>
-                    </div>
-                </div>
-            </t>
-        </xpath>
-    </t>
-</templates>
